+++ src/i18n_check/check/repeat_values.py
@@ -20,7 +20,7 @@ from rich import print as rprint
 from i18n_check.check.key_naming import audit_invalid_i18n_key_names, map_keys_to_files
 from i18n_check.utils import (
     config_i18n_src_file,
-    config_invalid_key_regexes_to_ignore,
+    config_repeat_values_regexes_to_ignore,
     config_src_directory,
     lower_and_remove_punctuation,
     read_json_file,
@@ -110,7 +110,7 @@ def analyze_and_generate_repeat_value_report(
                 key_file_dict={
                     "repeat_key": v for k, v in repeat_values_key_file_dict.items()
                 },
-                keys_to_ignore_regex=config_invalid_key_regexes_to_ignore,
+                keys_to_ignore_regex=config_repeat_values_regexes_to_ignore,
             )
 
             # Remove dummy value and add 'content_reference' for user to replace.